Add files via upload
This Python program calculates the most efficient way to represent a given monetary amount using dollars, quarters, dimes, nickels, and pennies. It takes an input from the user in the form of a float representing a dollar amount (with two decimal places), then calculates the minimum number of coins and bills required to make that amount.